+++ NEWS.md
@@ -1,5 +1,5 @@
 # 0.1.0.9000
-- wrap new functions for tidyr::gather() and tidyr::spread() (thanks @WilDoane)
+- added supoort for tidyr functions: gather, spread (thanks @WilDoane), drop_na (@jackhannah95), fill and replace_na
 - use clisymbols for ellipsis
 - add number of remaining rows to filter (#23)
 - bugfix: do not report negative NA (#18)